+++ nens_auth_client/tests/test_logout_view.py
@@ -3,14 +3,18 @@ from urllib.parse import urlparse
 from nens_auth_client import views
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser, User
+from django.core.exceptions import PermissionDenied
+
+import pytest
 
 
-def test_logout(rf, mocker, openid_configuration):
+@pytest.mark.parametrize("logged_in", [True, False])
+def test_logout(rf, mocker, openid_configuration, logged_in):
     django_logout = mocker.patch("nens_auth_client.views.django_auth.logout")
 
     request = rf.get("http://testserver/logout/?next=/a")
     request.session = {}
-    request.user = User()  # user is logged in initially
+    request.user = User() if logged_in else AnonymousUser()
     response = views.logout(request)
 
     # login generated a redirect to the LOGOUT_URL
@@ -19,7 +23,7 @@ def test_logout(rf, mocker, openid_configuration):
     assert url[:3] == ("https", "authserver", "/logout")
     qs = parse_qs(url.query)
     assert qs["client_id"] == [settings.NENS_AUTH_CLIENT_ID]
-    assert qs["logout_uri"] == ["http://testserver/logout/"]
+    assert qs["logout_uri"] == ["http://testserver/logout-success/"]
 
     # django logout was called
     assert django_logout.called
@@ -31,71 +35,48 @@ def test_logout(rf, mocker, openid_configuration):
     assert response._headers["cache-control"] == ("Cache-Control", "no-store")
 
 
-def test_logout_as_callback(rf, mocker):
-    django_logout = mocker.patch("nens_auth_client.views.django_auth.logout")
-
-    request = rf.get("http://testserver/logout/?next=/a")
-    request.session = {views.LOGOUT_REDIRECT_SESSION_KEY: "/b"}
-    request.user = AnonymousUser()  # user is not logged in anymore
-    response = views.logout(request)
-
-    # logout generated a redirect to the url stored in the session
-    assert response.status_code == 302
-    assert response.url == "/b"
-
-    # django logout was not called
-    assert not django_logout.called
-
-    # check if Cache-Control header is set to "no-store"
-    assert response._headers["cache-control"] == ("Cache-Control", "no-store")
-
-
-def test_logout_not_logged_in(rf, mocker):
-    # A user logs out without being logged in
-    django_logout = mocker.patch("nens_auth_client.views.django_auth.logout")
-
-    request = rf.get("http://testserver/logout/?next=/a")
-    request.session = {}
-    request.user = AnonymousUser()  # user is not logged in initially
-    response = views.logout(request)
-
-    # logout generated a redirect to the 'next' URL
-    assert response.status_code == 302
-    assert response.url == "/a"
-
-    # django logout was not called
-    assert not django_logout.called
-
-
 def test_logout_no_next_url(rf, mocker, openid_configuration):
     mocker.patch("nens_auth_client.views.django_auth.logout")
 
     request = rf.get("http://testserver/logout/")
     request.session = {}
-    request.user = User()  # user is logged in initially
     views.logout(request)
 
     # there is no redirect url stored in the session
     assert views.LOGOUT_REDIRECT_SESSION_KEY not in request.session
 
 
-def test_logout_as_callback_empty_session(rf, mocker, openid_configuration):
-    request = rf.get("http://testserver/logout/")
+def test_logout_success(rf, mocker):
+    request = rf.get("http://testserver/logout-success/")
+    request.session = {views.LOGOUT_REDIRECT_SESSION_KEY: "/b"}
+    request.user = AnonymousUser()  # user is not logged in anymore
+    response = views.logout_success(request)
+
+    # logout generated a redirect to the url stored in the session
+    assert response.status_code == 302
+    assert response.url == "/b"
+
+    # check if Cache-Control header is set to "no-store"
+    assert response._headers["cache-control"] == ("Cache-Control", "no-store")
+
+
+def test_logout_success_empty_session(rf, mocker, openid_configuration):
+    request = rf.get("http://testserver/logout-success/")
     request.session = {}
     request.user = AnonymousUser()  # user is not logged in anymore
-    response = views.logout(request)
+    response = views.logout_success(request)
 
     # logout generated a redirect to the default logout url
     assert response.status_code == 302
     assert response.url == settings.NENS_AUTH_DEFAULT_LOGOUT_URL
 
 
-def test_logout_not_logged_in_no_next_url(rf, mocker):
-    request = rf.get("http://testserver/logout/")
+def test_logout_success_logged_in(rf, mocker):
+    # This should only be possible if the user typed in this URL himself.
+    # Give PermissionDenied.
+    request = rf.get("http://testserver/logout-success/")
     request.session = {}
-    request.user = AnonymousUser()  # user is not logged in initially
-    response = views.logout(request)
+    request.user = User()  # user is logged in
 
-    # logout generated a redirect to the default logout url
-    assert response.status_code == 302
-    assert response.url == settings.NENS_AUTH_DEFAULT_LOGOUT_URL
+    with pytest.raises(PermissionDenied):
+        views.logout_success(request)