+++ nens_auth_client/urls.py
@@ -62,6 +62,7 @@ urlpatterns = [
     re_path("^authorize/$", views.authorize, name="authorize"),
     re_path("^login/$", views.login, name="login"),
     re_path("^logout/$", views.logout, name="logout"),
+    re_path("^logout-success/$", views.logout_success, name="logout-success"),
     re_path(
         r"^invitations/(?P<slug>\w+)/accept/",
         views.accept_invitation,