+++ nens_auth_client/views.py
@@ -52,6 +52,7 @@ def login(request):
         settings.NENS_AUTH_DEFAULT_SUCCESS_URL
       invitation: an optional Invitation id. On authorization success, a user will be
         created and permissions from the Invitation are applied.
+      force_logout: if "true", force local and remote logout
 
     Response:
       HTTP 302 Redirect to AWS Cognito (according to the OpenID Connect standard)
@@ -66,8 +67,12 @@ def login(request):
     # Get the success redirect url
     success_url = _get_redirect_from_next(request)
 
-    # If the user was already authenticated, redirect to the success url
-    if request.user.is_authenticated:
+    # Whether to force logout or not
+    force_logout = request.GET.get("force_logout") == "true"
+    if force_logout:
+        django_auth.logout(request)
+    elif request.user.is_authenticated:
+        # If the user was already authenticated, redirect to the success url
         return HttpResponseRedirect(
             success_url or settings.NENS_AUTH_DEFAULT_SUCCESS_URL
         )
@@ -86,7 +91,11 @@ def login(request):
     redirect_uri = request.build_absolute_uri(
         reverse(settings.NENS_AUTH_URL_NAMESPACE + "authorize")
     )
-    return client.authorize_redirect(request, redirect_uri)
+
+    if force_logout:
+        return client.logout_redirect(request, redirect_uri, login_after=True)
+    else:
+        return client.authorize_redirect(request, redirect_uri)
 
 
 @cache_control(no_store=True)
@@ -162,31 +171,17 @@ def authorize(request):
 def logout(request):
     """Logout the user (locally and remotely)
 
-    See the README for a description of the logout flow. Note that this view
-    is called twice in this flow.
+    See the README for a description of the logout flow.
 
     Query parameters:
-      next: the URL to redirect to after logout. If absolute, it
+      next: the URL to redirect to after logout-success. If absolute, it
         must match the domain of this request. Optional, default is set by
         settings.NENS_AUTH_DEFAULT_LOGOUT_URL
 
     Response:
     - if user is logged in locally: HTTP 302 Redirect to the remote logout URL.
       The user is logged out and the 'next' query parameter is stored in the session.
-    - if user is logged out locally: HTTP 302 Redirect to the 'next' query parameter
     """
-    if not request.user.is_authenticated:
-        # We are in step 3. (user is already logged out)
-        redirect_url = request.session.get(LOGOUT_REDIRECT_SESSION_KEY, None)
-        if redirect_url is None:
-            # If there is nothing in the session, the user called /logout
-            # without being logged in in the first place. Just use the 'next'
-            # parameter (or the default logout url).
-            redirect_url = _get_redirect_from_next(request)
-        return HttpResponseRedirect(
-            redirect_url or settings.NENS_AUTH_DEFAULT_LOGOUT_URL
-        )
-
     # Log the user out
     django_auth.logout(request)
 
@@ -197,10 +192,36 @@ def logout(request):
 
     # Redirect to authorization server
     logout_uri = request.build_absolute_uri(
-        reverse(settings.NENS_AUTH_URL_NAMESPACE + "logout")
+        reverse(settings.NENS_AUTH_URL_NAMESPACE + "logout-success")
     )
     client = get_oauth_client()
     return client.logout_redirect(request, logout_uri)
+
+
+@cache_control(no_store=True)
+def logout_success(request):
+    """Callback url for logout.
+
+    See the README for a description of the logout flow.
+
+    Query parameters:
+      next: the URL to redirect to after logout. If absolute, it
+        must match the domain of this request. Optional, default is set by
+        settings.NENS_AUTH_DEFAULT_LOGOUT_URL
+
+    Response:
+    - if user is logged out locally: HTTP 302 Redirect to the 'next' query parameter
+      that is stored in the session (or default logout url if unavailable).
+    - if user is logged in locally: HTTP 302 Redirect to the logout view
+    """
+    # If a user is still authenticated: this should not happen. Return 403.
+    if request.user.is_authenticated:
+        raise PermissionDenied("Logout failure")
+
+    redirect_url = request.session.get(
+        LOGOUT_REDIRECT_SESSION_KEY, settings.NENS_AUTH_DEFAULT_LOGOUT_URL
+    )
+    return HttpResponseRedirect(redirect_url)
 
 
 @cache_control(no_store=True)